requirements.txt
discord.py
python-dotenv~=1.1.1
app~=0.0.1